Renaming Wizard State pages
