+++ packages/client/src/pages/SignUp/wizards/MentorWizard/forms/MentorStep4.js
@@ -1,16 +1,10 @@
 import React from 'react';
-import styled from 'styled-components';
 import Button from '../../../../../components/Button';
 import Checkbox from '@material-ui/core/Checkbox';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import FormGroup from '@material-ui/core/FormGroup';
 
-const WizardInput = styled.div`
-    margin-bottom: 1em;
-    min-width: 120px;
-`;
-
-const FourthPageForm = (props) => {
+const MentorStep4 = (props) => {
 
     const termsOfServiceCheckbox = (
         <Checkbox
@@ -60,4 +54,4 @@ const FourthPageForm = (props) => {
 
 }
 
-export default FourthPageForm;
+export default MentorStep4;
