+++ packages/client/src/pages/SignUp/wizards/ParentWizard/index.js
@@ -3,10 +3,10 @@ import styled from 'styled-components';
 import Wizard from '../../../../components/Wizard';
 import Button from '../../../../components/Button';
 
-import FirstPageForm from './forms/FirstPageForm.js';
-import SecondPageForm from './forms/SecondPageForm.js';
-import ThirdPageForm from './forms/ThirdPageForm.js';
-import FourthPageForm from './forms/FourthPageForm.js';
+import ParentStep1 from './forms/ParentStep1.js';
+import ParentStep2 from './forms/ParentStep2.js';
+import ParentStep3 from './forms/ParentStep3.js';
+import ParentStep4 from './forms/ParentStep4.js';
 
 const SignUpChildWrapper = styled.div`
     display: flex;
@@ -61,99 +61,97 @@ const updateRegisteredChild = (index, data) => {
     parentWizardSignUpData.registeredChildren[index] = mergedChildRegistration;
 }
 
-const ParentWizard = () => {
+const FirstPage = () => {
 
-    const FirstPage = () => {
+    const [state, setState] = useState({});
 
-        const [state, setState] = useState({});
+    const handleChange = (event) => {
+        setState({ ...state, [event.target.name]: event.target.value });
+        updateParentWizardSignUpData({ [event.target.name]: event.target.value });
+    };
 
-        const handleChange = (event) => {
-            setState({ ...state, [event.target.name]: event.target.value });
-            updateParentWizardSignUpData({ [event.target.name]: event.target.value });
-        };
+    return (
+        <SignUpChildWrapper>
+            <ParentStep1 data={parentWizardSignUpData} handleChange={handleChange} />
+        </SignUpChildWrapper>
+    );
+}
 
-        return (
-            <SignUpChildWrapper>
-                <FirstPageForm data={parentWizardSignUpData} handleChange={handleChange} />
-            </SignUpChildWrapper>
-        );
-    }
+const SecondPage = () => {
 
-    const SecondPage = () => {
+    const [state, setState] = useState({});
 
-        const [state, setState] = useState({});
+    let children = parentWizardSignUpData.registeredChildren.map((item, index) => {
+        return <ParentStep2 data={parentWizardSignUpData} index={index} updateRegisteredChild={updateRegisteredChild} />;
+    });
 
-        let children = parentWizardSignUpData.registeredChildren.map((item, index) => {
-            return <SecondPageForm data={parentWizardSignUpData} index={index} updateRegisteredChild={updateRegisteredChild} />;
+    const handleAddClick = (event) => {
+        event.preventDefault();
+        parentWizardSignUpData.registeredChildren.push({
+            selectedSubjects: [],
+            gradeLevel: '',
         });
+        setState({ ...state });
+    }
 
-        const handleAddClick = (event) => {
-            event.preventDefault();
-            parentWizardSignUpData.registeredChildren.push({
-                selectedSubjects: [],
-                gradeLevel: '',
-            });
-            setState({ ...state });
-        }
+    const handleRemoveClick = (event) => {
+        event.preventDefault();
+        parentWizardSignUpData.registeredChildren.pop();
+        setState({ ...state });
+    }
 
-        const handleRemoveClick = (event) => {
-            event.preventDefault();
-            parentWizardSignUpData.registeredChildren.pop();
-            setState({ ...state });
-        }
+    const showRemoveChildButton = parentWizardSignUpData.registeredChildren.length > 1;
 
-        const showRemoveChildButton = parentWizardSignUpData.registeredChildren.length > 1;
-
-        return (
-            <SignUpChildWrapper>
-                {children}
-                <ChildSignUpButtonWrapper>
-                    <Button onClick={handleAddClick}>
-                        Add Child
+    return (
+        <SignUpChildWrapper>
+            {children}
+            <ChildSignUpButtonWrapper>
+                <Button onClick={handleAddClick}>
+                    Add Child
                     </Button>
-                    {showRemoveChildButton ?
-                        <Button onClick={handleRemoveClick}>
-                            Remove Child
+                {showRemoveChildButton ?
+                    <Button onClick={handleRemoveClick}>
+                        Remove Child
                         </Button> : null}
-                </ChildSignUpButtonWrapper>
-            </SignUpChildWrapper>
-        );
-    }
+            </ChildSignUpButtonWrapper>
+        </SignUpChildWrapper>
+    );
+}
 
-    const ThirdPage = () => {
+const ThirdPage = () => {
 
-        const [state, setState] = useState({});
+    const [state, setState] = useState({});
 
-        const handleChange = (event) => {
-            setState({ ...state, [event.target.name]: event.target.value });
-            updateParentWizardSignUpData({ [event.target.name]: event.target.value });
-        };
+    const handleChange = (event) => {
+        setState({ ...state, [event.target.name]: event.target.value });
+        updateParentWizardSignUpData({ [event.target.name]: event.target.value });
+    };
 
-        return (
-            <SignUpChildWrapper>
-                <ThirdPageForm data={parentWizardSignUpData} handleChange={handleChange}/>
-            </SignUpChildWrapper>
-        );
-    }
+    return (
+        <SignUpChildWrapper>
+            <ParentStep3 data={parentWizardSignUpData} handleChange={handleChange} />
+        </SignUpChildWrapper>
+    );
+}
 
-    const FourthPage = () => {
+const FourthPage = () => {
 
-        const [state, setState] = useState({});
+    const [state, setState] = useState({});
 
-        const handleCheck = (event) => {
-            setState({ ...state, [event.target.name]: event.target.checked });
-            updateParentWizardSignUpData({ [event.target.name]: event.target.checked });
-        };
-        
-        return (
-            <SignUpChildWrapper>
-                <FourthPageForm data={parentWizardSignUpData} handleCheck={handleCheck}/>
-            </SignUpChildWrapper>
-        );
-    }
+    const handleCheck = (event) => {
+        setState({ ...state, [event.target.name]: event.target.checked });
+        updateParentWizardSignUpData({ [event.target.name]: event.target.checked });
+    };
 
-    const children = [<FirstPage />, <SecondPage />, <ThirdPage />, <FourthPage />];
-    
+    return (
+        <SignUpChildWrapper>
+            <ParentStep4 data={parentWizardSignUpData} handleCheck={handleCheck} />
+        </SignUpChildWrapper>
+    );
+}
+
+const ParentWizard = () => {
+    const children = [<FirstPage />, <SecondPage />, <ThirdPage />, <FourthPage />];    
     return <Wizard content={children} />;
 }
